requirements.txt
Flask
moviepy
pydub
gtts
pillow==9.5.0
